:heavy_plus_sign: point db list get func add, point member register func add

+++ rest_server/config/server_config.go
@@ -34,7 +34,7 @@ type ApiAuth struct {
 }
 
 type MssqlPoint struct {
-	DBID     int    `json:"db_id" yaml:"db_id"`
+	DBID     int64  `json:"db_id" yaml:"db_id"`
 	Host     string `json:"host" yaml:"host"`
 	Port     string `json:"port" yaml:"port"`
 	ID       string `json:"id" yaml:"id"`
